requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
